chore: refactored chain token selector
